enabling this by default leads to ordering issues
specifically, Plack::Middleware::Session needs to be on the outside of
Plack::Middleware::HTTPExceptions, because ::Session needs access to the
finalized response in order to add things to the cookie or whatever it's
using for state. Most other middleware needs to be on the inside of
::HTTPExceptions, so that it can catch exceptions thrown by those other
middleware. trying to handle both cases would require too much magic, so
just don't do it at all for now.