fix handling of routes with equivalent paths
this makes routes where the paths are equivalent work properly (they can
be differentiated with validations, for instance). it does break
overridden inherited routes a bit, but i think that's a less commonly
used feature, so i'd rather have that broken for now. i know how to fix
that aspect too, but it'll require a decent bit of work, so later.